feat: use imperative mapping
BREAKING CHANGE: we were using declarative mapping which, AFAIK, does not allow for any customization of the tables that will be created. This commit migrates to imperative mapping with default tables that are identical to what they were before, but does not require the user to actually use those tables exactly. If they so desire, they might create tables with different column names or types (e.g. JSONB instead of JSON for Postgres), add indexes, etc, and simply map them later to the Python classes that represent those tables in memory.